+++ services/celest_cloud_auth/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.0+2
 
 - fix: Ensure Cedar types are registered during seeding
+- fix: Ensure session is linked to user cork
 
 ## 0.1.0+1